[core] First batch of unchecked integrations loading
