+++ src/DDTrace/Integrations/Curl/CurlIntegration.php
@@ -11,7 +11,6 @@ use DDTrace\Tag;
 use DDTrace\Type;
 use DDTrace\Util\ArrayKVStore;
 use DDTrace\GlobalTracer;
-use DDTrace\Util\Versions;
 
 /**
  * Integration for curl php client.
@@ -25,8 +24,8 @@ class CurlIntegration
      */
     public static function load()
     {
-        if (!function_exists('curl_exec')) {
-            // `curl_exec` doesn't come from an autoloader, if it does not exists we can return this integration as
+        if (!extension_loaded('curl')) {
+            // `curl` extension is not loaded, if it does not exists we can return this integration as
             // not available.
             return Integration::NOT_AVAILABLE;
         }
