+++ src/DDTrace/Integrations/Integration.php
@@ -21,9 +21,6 @@ abstract class Integration
 
     public static function load()
     {
-        if (!class_exists(static::CLASS_NAME)) {
-            return self::NOT_LOADED;
-        }
         // See comment on the commented out abstract function definition.
         static::loadIntegration();
