+++ src/DDTrace/Integrations/Memcached/MemcachedIntegration.php
@@ -29,7 +29,7 @@ class MemcachedIntegration
 
     public static function load()
     {
-        if (!class_exists('Memcached')) {
+        if (!extension_loaded('memcached')) {
             // Memcached is provided through an extension and not through a class loader.
             return Integration::NOT_AVAILABLE;
         }
